refactored gui creation
- broke it into classes and methods for readability and maintainability
