+++ src/main/java/mod/icy_turtle/friendhighlighter/util/MultiPartGUIElement.java
@@ -1,5 +1,6 @@
 package mod.icy_turtle.friendhighlighter.util;
 
+import me.shedaniel.clothconfig2.api.AbstractConfigListEntry;
 import me.shedaniel.clothconfig2.gui.entries.MultiElementListEntry;
 import net.minecraft.text.Text;
 
@@ -7,7 +8,7 @@ import java.util.List;
 
 public class MultiPartGUIElement<T> extends MultiElementListEntry<T>
 {
-	public MultiPartGUIElement(Text categoryName, T object, List list, boolean defaultExpanded)
+	public MultiPartGUIElement(Text categoryName, T object, List<AbstractConfigListEntry<?>> list, boolean defaultExpanded)
 	{
 		super(categoryName, object, list, defaultExpanded);
 	}
